+++ sdk/resourcemanager/testbase/armtestbase/availableos_client.go
@@ -47,7 +47,7 @@ func NewAvailableOSClient(subscriptionID string, credential azcore.TokenCredenti
 // Get - Gets an available OS to run a package under a Test Base Account.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2020-12-16-preview
+// Generated from API version 2022-04-01-preview
 //   - resourceGroupName - The name of the resource group that contains the resource.
 //   - testBaseAccountName - The resource name of the Test Base Account.
 //   - availableOSResourceName - The resource name of an Available OS.
@@ -91,7 +91,7 @@ func (client *AvailableOSClient) getCreateRequest(ctx context.Context, resourceG
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2020-12-16-preview")
+	reqQP.Set("api-version", "2022-04-01-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -108,7 +108,7 @@ func (client *AvailableOSClient) getHandleResponse(resp *http.Response) (Availab
 
 // NewListPager - Lists all the available OSs to run a package under a Test Base Account.
 //
-// Generated from API version 2020-12-16-preview
+// Generated from API version 2022-04-01-preview
 //   - resourceGroupName - The name of the resource group that contains the resource.
 //   - testBaseAccountName - The resource name of the Test Base Account.
 //   - osUpdateType - The type of the OS Update.
@@ -162,7 +162,7 @@ func (client *AvailableOSClient) listCreateRequest(ctx context.Context, resource
 	}
 	reqQP := req.Raw().URL.Query()
 	reqQP.Set("osUpdateType", string(osUpdateType))
-	reqQP.Set("api-version", "2020-12-16-preview")
+	reqQP.Set("api-version", "2022-04-01-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil