+++ sdk/resourcemanager/testbase/armtestbase/constants.go
@@ -11,7 +11,7 @@ package armtestbase
 
 const (
 	moduleName    = "armtestbase"
-	moduleVersion = "v0.6.1"
+	moduleVersion = "v0.7.0"
 )
 
 // Action - The action of the command.
@@ -221,6 +221,7 @@ func PossibleGradeValues() []Grade {
 	}
 }
 
+// OsUpdateType - Specifies the OS update type to test against.
 type OsUpdateType string
 
 const (