+++ sdk/resourcemanager/testbase/armtestbase/osupdates_client.go
@@ -47,7 +47,7 @@ func NewOSUpdatesClient(subscriptionID string, credential azcore.TokenCredential
 // Get - Gets an OS Update by name in which the package was tested before.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2020-12-16-preview
+// Generated from API version 2022-04-01-preview
 //   - resourceGroupName - The name of the resource group that contains the resource.
 //   - testBaseAccountName - The resource name of the Test Base Account.
 //   - packageName - The resource name of the Test Base Package.
@@ -96,7 +96,7 @@ func (client *OSUpdatesClient) getCreateRequest(ctx context.Context, resourceGro
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2020-12-16-preview")
+	reqQP.Set("api-version", "2022-04-01-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -113,7 +113,7 @@ func (client *OSUpdatesClient) getHandleResponse(resp *http.Response) (OSUpdates
 
 // NewListPager - Lists the OS Updates in which the package were tested before.
 //
-// Generated from API version 2020-12-16-preview
+// Generated from API version 2022-04-01-preview
 //   - resourceGroupName - The name of the resource group that contains the resource.
 //   - testBaseAccountName - The resource name of the Test Base Account.
 //   - packageName - The resource name of the Test Base Package.
@@ -172,7 +172,7 @@ func (client *OSUpdatesClient) listCreateRequest(ctx context.Context, resourceGr
 	}
 	reqQP := req.Raw().URL.Query()
 	reqQP.Set("osUpdateType", string(osUpdateType))
-	reqQP.Set("api-version", "2020-12-16-preview")
+	reqQP.Set("api-version", "2022-04-01-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil