+++ sdk/resourcemanager/testbase/armtestbase/CHANGELOG.md
@@ -1,5 +1,32 @@
 # Release History
 
+## 0.7.0 (2023-06-07)
+### Features Added
+
+- New function `NewBillingHubServiceClient(string, azcore.TokenCredential, *arm.ClientOptions) (*BillingHubServiceClient, error)`
+- New function `*BillingHubServiceClient.GetFreeHourBalance(context.Context, string, string, *BillingHubServiceClientGetFreeHourBalanceOptions) (BillingHubServiceClientGetFreeHourBalanceResponse, error)`
+- New function `*BillingHubServiceClient.GetUsage(context.Context, string, string, *BillingHubServiceClientGetUsageOptions) (BillingHubServiceClientGetUsageResponse, error)`
+- New function `*ClientFactory.NewBillingHubServiceClient() *BillingHubServiceClient`
+- New function `*PackagesClient.RunTest(context.Context, string, string, string, *PackagesClientRunTestOptions) (PackagesClientRunTestResponse, error)`
+- New function `*TestResultsClient.GetConsoleLogDownloadURL(context.Context, string, string, string, string, TestResultConsoleLogDownloadURLParameters, *TestResultsClientGetConsoleLogDownloadURLOptions) (TestResultsClientGetConsoleLogDownloadURLResponse, error)`
+- New struct `BillingHubExecutionUsageDetail`
+- New struct `BillingHubFreeHourIncrementEntry`
+- New struct `BillingHubGetFreeHourBalanceResponse`
+- New struct `BillingHubGetUsageRequest`
+- New struct `BillingHubGetUsageResponse`
+- New struct `BillingHubPackageUsage`
+- New struct `BillingHubUsage`
+- New struct `BillingHubUsageGroup`
+- New struct `BillingHubUsageGroupedByUpdateType`
+- New struct `PackageRunTestParameters`
+- New struct `TestResultConsoleLogDownloadURLParameters`
+- New field `StderrLogFileName`, `StdoutLogFileName` in struct `ScriptExecutionResult`
+- New field `BaselineOSs` in struct `TargetOSInfo`
+- New field `ValidationResultID` in struct `Test`
+- New field `InteropMediaType`, `InteropMediaVersion` in struct `TestResultProperties`
+- New field `PackageTags` in struct `TestSummaryProperties`
+
+
 ## 0.6.1 (2023-04-14)
 ### Bug Fixes